backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0
redis>=5.0.0
celery>=5.4.0

# LLM & Agent
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0

# Utilities
httpx>=0.25.0
python-multipart>=0.0.9
python-docx>=1.1.0
PyJWT>=2.10.0
pwdlib[argon2]>=0.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
